requirements.txt
python-dotenv
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2
cryptography
python-multipart
python-jose[cryptography]
passlib[bcrypt]

